requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
plotly>=5.18.0
pydantic>=2.5.0
pyyaml>=6.0.0
joblib>=1.3.0
pytest>=7.4.0
httpx>=0.25.0
